StudyMind/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
endee
sentence-transformers==3.0.1
pypdf==4.3.1
openai==1.40.0
groq==0.9.0
python-dotenv==1.0.1
